[#6] Develop TableJoin Library
* Renamed TableJoinClientTests to TableJoinTest.
